refactor(vendingmachine.service.*) : VendingMachineService에서 SettingCoinService 분리
